my_photo_album_2/pyproject.toml
[tool.poetry]
name = "my_photo_album_2"
version = "0.1.0"
description = "Personal face-based photo classification and search system."
authors = ["you@example.com"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
opencv-python-headless = ">=4.8.0"
numpy = "*"
matplotlib = "*"
fastapi = ">=0.100.0"
uvicorn = { extras = ["standard"], version = ">=0.24.0" }
python-multipart = ">=0.0.6"
Pillow = ">=9.0.0"
requests = "*"
PyYAML = "*"
scikit-learn = "*"
mediapipe = "*"
face-recognition = "^1.3.0"
jinja2 = "^3.1.6"

[tool.poetry.group.dev.dependencies]
ruff = ">=0.4.0"
pytest = ">=7.4.0"
mypy = ">=1.8.0"
pre-commit = ">=3.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
